Move SnapshotConfig into its own module
Also move ArchiveFormat to snapshot_utils
